All files moved into new folder structure for profile/ hub/ project/
