+++ src/components/hub/HubsList.tsx
@@ -2,7 +2,7 @@ import {List} from "grommet";
 import {User} from "@supabase/auth-helpers-react";
 import {useRouter} from "next/router";
 
-import {Hub} from "../utils/types";
+import {Hub} from "../../utils/types";
 
 type Props = {
     hubs: Array<Hub>
