+++ src/components/hub/NewHubStep1.tsx
@@ -5,36 +5,36 @@ import {
     FormField, Paragraph,
     TextInput,
 } from 'grommet'
-import { useState } from 'react'
-import {Session, useSession, useSupabaseClient, useUser} from "@supabase/auth-helpers-react";
+import {useState} from 'react'
+import {Session, useSession, useSupabaseClient} from "@supabase/auth-helpers-react";
 
-import { Database } from "../utils/database.types";
-import {Project} from "../utils/types";
+import {Database} from "../../utils/database.types";
+import {Hub} from "../../utils/types";
 import {useRouter} from "next/router";
 
-export default function NewProjectStep1() {
+export default function NewHubStep1() {
     const supabase = useSupabaseClient<Database>()
     const session = useSession()
     const router = useRouter()
 
-    const [name, setName] = useState('')
-    const [description, setDescription] = useState('')
+    const [hubName, setName] = useState('')
+    const [hubDescription, setDescription] = useState('')
     const [loading, setLoading] = useState(false)
 
     if (!session) return (<Paragraph>Unauthorized</Paragraph>)
 
-    async function createProject(session: Session) {
+    async function createHub(session: Session) {
         try {
             setLoading(true)
 
-            const { data, error } = await supabase.rpc('add_project', { name: name, description: description, firstadmin: session.user.id })
+            const { data, error } = await supabase.rpc('add_hub', { name: hubName, description: hubDescription, firstadmin: session.user.id })
 
             if (error)
                 throw error
-            alert('Project created with ID '+data)
+            alert('Hub created with ID '+data)
             return data
         } catch (error) {
-            alert('Error creating the project! Message: '+JSON.stringify(error))
+            alert('Error creating the hub! Message: '+JSON.stringify(error))
             console.log(error)
         } finally {
             setLoading(false)
@@ -43,16 +43,16 @@ export default function NewProjectStep1() {
 
     return (
         <Box width="large" elevation="medium" round pad="large">
-            <Form<Project>
+            <Form<Hub>
                 onSubmit={async () => {
-                    const projectId = await createProject(session)
-                    router.push("/project/"+projectId)
+                    const hubId = await createHub(session)
+                    router.push("/hub/"+hubId)
                 }}>
                     <FormField width="100%" name="name" htmlFor="nameId" label="Name" required>
-                        <TextInput id="nameId" name="name" type="name" value={name} onChange={event => setName(event.target.value)}/>
+                        <TextInput id="nameId" name="name" type="name" value={hubName} onChange={event => setName(event.target.value)}/>
                     </FormField>
                     <FormField width="100%" name="description" htmlFor="descriptionId" label="Description" required>
-                        <TextInput id="descriptionId" name="description" type="text" value={description} onChange={event => setDescription(event.target.value)}/>
+                        <TextInput id="descriptionId" name="description" type="text" value={hubDescription} onChange={event => setDescription(event.target.value)}/>
                     </FormField>
                 <Box direction="row" gap="medium" width="50%" margin={{ horizontal: 'auto', top: 'large' }}>
                     <Button type="submit" primary label={loading ? 'Loading ...' : 'Submit'} disabled={loading} />
