+++ src/components/project/ProjectsList.tsx
@@ -1,7 +1,7 @@
 import {List} from "grommet";
 import {useRouter} from "next/router";
 
-import {Project} from "../utils/types";
+import {Project} from "../../utils/types";
 
 type Props = {
     projects: Array<Project>
