+++ src/pages/hub/new.tsx
@@ -1,6 +1,6 @@
 
 import {useSession} from "@supabase/auth-helpers-react";
-import NewHubStep1 from "../../components/NewHubStep1";
+import NewHubStep1 from "../../components/hub/NewHubStep1";
 import {Paragraph} from "grommet";
 
 export default function NewHub() {
